Extract download file function
