+++ airbyte-webapp/src/utils/file.ts
@@ -1,0 +1,8 @@
+export const downloadFile = (blob: Blob, name: string) => {
+  const element = document.createElement("a");
+  element.href = URL.createObjectURL(blob);
+  element.download = name;
+  document.body.appendChild(element); // Required for this to work in FireFox
+  element.click();
+  document.body.removeChild(element);
+};
